fixup! fixup! fixup! DTLS 1.3 Update Epoch to a 64 bit counter and don't allow wrapping

+++ include/internal/recordmethod.h
@@ -229,7 +229,7 @@ struct ossl_record_method_st {
      */
     int (*read_record)(OSSL_RECORD_LAYER *rl, void **rechandle, int *rversion,
         uint8_t *type, const unsigned char **data, size_t *datalen,
-        uint16_t *epoch, uint64_t *seq_num);
+        uint64_t *epoch, uint64_t *seq_num);
     /*
      * Release length bytes from a buffer associated with a record previously
      * read with read_record. Once all the bytes from a record are released, the
